Fixed listing feature image click issue.
Whenever you add a featured image of a listing it stopped slider and shows the image only. Previously that image was unclickable. Currently, you can click on it and it will go the inner or single page of it. If you don't want to set a single image just ignore the feature image and add the image gallery section of the editor page.